Defend against over-large stderr files
